Fix a perhaps-incorrect typedef
This type is being used in several printf lines as a char, but it's
typedef'd to a 32-bit int. I'm not sure which direction is "wrong."

+++ c10/hammerblade/emul/bsg_manycore_coordinate.h
@@ -50,7 +50,7 @@ extern "C" {
                 HB_MC_COORDINATE_MAX = 32
         } hb_mc_coordinate_limit_t;
 
-        typedef uint32_t hb_mc_idx_t;
+        typedef uint8_t hb_mc_idx_t;
 
         typedef struct __hb_mc_coordinate_t{
                 hb_mc_idx_t x;
